HW-5.2.2 refactoring
