+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/entity/auth/AuthUserEntity.java
@@ -1,6 +1,6 @@
 package guru.qa.niffler.data.entity.auth;
 
-import guru.qa.niffler.model.AuthUserJson;
+import guru.qa.niffler.model.auth.AuthUserJson;
 import java.io.Serializable;
 import java.util.UUID;
 import lombok.Getter;
