+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/entity/spend/CategoryEntity.java
@@ -1,6 +1,6 @@
 package guru.qa.niffler.data.entity.spend;
 
-import guru.qa.niffler.model.CategoryJson;
+import guru.qa.niffler.model.spend.CategoryJson;
 import java.io.Serializable;
 import java.util.UUID;
 import lombok.Getter;
