+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/entity/spend/SpendEntity.java
@@ -1,7 +1,7 @@
 package guru.qa.niffler.data.entity.spend;
 
-import guru.qa.niffler.model.CurrencyValues;
-import guru.qa.niffler.model.SpendJson;
+import guru.qa.niffler.model.spend.CurrencyValues;
+import guru.qa.niffler.model.spend.SpendJson;
 import java.io.Serializable;
 import java.sql.Date;
 import java.util.UUID;
