+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/entity/userdata/UserEntity.java
@@ -1,7 +1,7 @@
 package guru.qa.niffler.data.entity.userdata;
 
-import guru.qa.niffler.model.CurrencyValues;
-import guru.qa.niffler.model.UserJson;
+import guru.qa.niffler.model.spend.CurrencyValues;
+import guru.qa.niffler.model.user.UserJson;
 import java.io.Serializable;
 import java.util.UUID;
 import lombok.Getter;
