+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/impl/userdata/UserdataUserDaoJdbc.java
@@ -4,7 +4,7 @@ import static guru.qa.niffler.data.tpl.Connections.holder;
 import guru.qa.niffler.config.Config;
 import guru.qa.niffler.data.dao.userdata.UserdataUserDao;
 import guru.qa.niffler.data.entity.userdata.UserEntity;
-import guru.qa.niffler.model.CurrencyValues;
+import guru.qa.niffler.model.spend.CurrencyValues;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
