+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/mapper/spend/SpendEntityRowMapper.java
@@ -2,7 +2,7 @@ package guru.qa.niffler.data.mapper.spend;
 
 import guru.qa.niffler.data.entity.spend.CategoryEntity;
 import guru.qa.niffler.data.entity.spend.SpendEntity;
-import guru.qa.niffler.model.CurrencyValues;
+import guru.qa.niffler.model.spend.CurrencyValues;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.UUID;
