+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extension/CategoryExtension.java
@@ -3,7 +3,7 @@ package guru.qa.niffler.jupiter.extension;
 import static utils.RandomDataUtils.randomCategoryName;
 import guru.qa.niffler.jupiter.annotation.Category;
 import guru.qa.niffler.jupiter.annotation.meta.User;
-import guru.qa.niffler.model.CategoryJson;
+import guru.qa.niffler.model.spend.CategoryJson;
 import guru.qa.niffler.service.SpendClient;
 import guru.qa.niffler.service.SpendDbClient;
 import java.util.Optional;
