+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extension/UserQueueExtension.java
@@ -2,7 +2,7 @@ package guru.qa.niffler.jupiter.extension;
 
 import guru.qa.niffler.jupiter.annotation.UserType;
 import guru.qa.niffler.jupiter.annotation.UserType.FriendType;
-import guru.qa.niffler.model.StaticUser;
+import guru.qa.niffler.model.user.StaticUser;
 import io.qameta.allure.Allure;
 import java.util.Arrays;
 import java.util.Date;
