+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/auth/AuthorityJson.java
@@ -1,4 +1,4 @@
-package guru.qa.niffler.model;
+package guru.qa.niffler.model.auth;
 
 import guru.qa.niffler.data.entity.auth.Authority;
 import guru.qa.niffler.data.entity.auth.AuthorityEntity;
