+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/spend/CategoryJson.java
@@ -1,4 +1,4 @@
-package guru.qa.niffler.model;
+package guru.qa.niffler.model.spend;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import guru.qa.niffler.data.entity.spend.CategoryEntity;
