+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/spend/CurrencyValues.java
@@ -1,4 +1,4 @@
-package guru.qa.niffler.model;
+package guru.qa.niffler.model.spend;
 
 public enum CurrencyValues {
   RUB, USD, EUR, KZT
