+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/user/StaticUser.java
@@ -1,4 +1,4 @@
-package guru.qa.niffler.model;
+package guru.qa.niffler.model.user;
 
 public record StaticUser(String username, String password, String friend, String income,
                          String outcome) {
