+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/service/SpendClient.java
@@ -1,7 +1,7 @@
 package guru.qa.niffler.service;
 
-import guru.qa.niffler.model.CategoryJson;
-import guru.qa.niffler.model.SpendJson;
+import guru.qa.niffler.model.spend.CategoryJson;
+import guru.qa.niffler.model.spend.SpendJson;
 import java.util.List;
 import java.util.Optional;
