+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/service/SpendDbClient.java
@@ -8,8 +8,8 @@ import guru.qa.niffler.data.entity.spend.SpendEntity;
 import guru.qa.niffler.data.impl.spend.CategoryDaoJdbc;
 import guru.qa.niffler.data.impl.spend.SpendDaoJdbc;
 import guru.qa.niffler.data.tpl.JdbcTransactionTemplate;
-import guru.qa.niffler.model.CategoryJson;
-import guru.qa.niffler.model.SpendJson;
+import guru.qa.niffler.model.spend.CategoryJson;
+import guru.qa.niffler.model.spend.SpendJson;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
