+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/service/UserClient.java
@@ -1,6 +1,6 @@
 package guru.qa.niffler.service;
 
-import guru.qa.niffler.model.UserJson;
+import guru.qa.niffler.model.user.UserJson;
 
 public interface UserClient {
