+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/service/UserDbClient.java
@@ -12,8 +12,8 @@ import guru.qa.niffler.data.impl.auth.AuthAuthorityDaoSpringJdbc;
 import guru.qa.niffler.data.impl.auth.AuthUserDaoSpringJdbc;
 import guru.qa.niffler.data.impl.userdata.UserdataUserDaoSpringJdbc;
 import guru.qa.niffler.data.tpl.XaTransactionTemplate;
-import guru.qa.niffler.model.AuthUserJson;
-import guru.qa.niffler.model.UserJson;
+import guru.qa.niffler.model.auth.AuthUserJson;
+import guru.qa.niffler.model.user.UserJson;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
