+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/FriendsWebTest.java
@@ -6,7 +6,7 @@ import guru.qa.niffler.jupiter.annotation.UserType;
 import guru.qa.niffler.jupiter.annotation.UserType.FriendType;
 import guru.qa.niffler.jupiter.extension.BrowserExtension;
 import guru.qa.niffler.jupiter.extension.UserQueueExtension;
-import guru.qa.niffler.model.StaticUser;
+import guru.qa.niffler.model.user.StaticUser;
 import guru.qa.niffler.page.LoginPage;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
