+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/TransactionTest.java
@@ -13,11 +13,11 @@ import guru.qa.niffler.data.impl.userdata.UserdataUserDaoJdbc;
 import guru.qa.niffler.data.impl.userdata.UserdataUserDaoSpringJdbc;
 import guru.qa.niffler.data.tpl.DataSources;
 import guru.qa.niffler.data.tpl.XaTransactionTemplate;
-import guru.qa.niffler.model.AuthUserJson;
-import guru.qa.niffler.model.CategoryJson;
-import guru.qa.niffler.model.CurrencyValues;
-import guru.qa.niffler.model.SpendJson;
-import guru.qa.niffler.model.UserJson;
+import guru.qa.niffler.model.auth.AuthUserJson;
+import guru.qa.niffler.model.spend.CategoryJson;
+import guru.qa.niffler.model.spend.CurrencyValues;
+import guru.qa.niffler.model.spend.SpendJson;
+import guru.qa.niffler.model.user.UserJson;
 import guru.qa.niffler.service.SpendDbClient;
 import guru.qa.niffler.service.UserDbClient;
 import java.sql.Connection;
